refactor : extract claimMechanic into its own file with math/effects split
Moved claimMechanic + the two mode resolvers out of mechanics/functions.php
into mechanics/claimMechanic.php, following the per-mechanic file pattern
(attackMechanic, investigateMechanic, etc.). The dispatcher now collects
resolution descriptors from claimByWorkerMath or claimByWorkerLeaderMath
and runs a single shared post-resolution loop for observer view-reports,
CKE writes, self-reports, and zones UPDATE. Behaviour preserved (345/345).
